+++ rgb-to-hex.js
@@ -1,0 +1,15 @@
+//RGB to Hex
+const componentToHex = (c) => {
+  var hex = c.toString(16);
+  return hex.length == 1 ? "0" + hex : hex;
+}
+
+const rgbToHex = (r, g, b) => {
+  return "#" + componentToHex(parseInt(r)) + componentToHex(parseInt(g)) + componentToHex(parseInt(b));
+}
+
+const dataRgb = {
+    componentToHex, rgbToHex
+}
+
+module.exports = dataRgb;
